Updating function "BuildUriForListRecords"
Found Error while retrieving records. on base of Debugging Removing "HttpUtility.UrlEncode function for TableName" on line no 278

Reason:
Its adding a + sign between the table name if it contains a space that puts an error when uri is made. If we leave it blank it does automatically put %20 as space.