Remove 'internal sets' to allow users to mock out responses.
Prior to this commit, it was non-trivial for end users of this Airtable
client to mock out responses to e.g. a list webhooks call.  Whilst they
could mock out a method to return a list of webhooks, they would not be
able to specify what properties should be set inside those webhooks due
to the 'internal set' modifier.

This commit removes all occurrences of the 'internal' keyword from each
Airtable object property so that users of the library can now have full
controll over mocking out exact responses that this client responds
with.